+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enable nullable reference types feature.
 
+### Removed
+
+- Remove obsolete `CliCommandException.Create` static method.
+
 ## [2.2.0] - 2022-07-29
 
 ### Added